Refactor language
