+++ src/features/aktuelt/Aktuelt.astro
@@ -5,7 +5,7 @@ import { getAudience, getOboToken } from "@utils/server/token";
 import { BodyShort, Skeleton } from "@navikt/ds-react";
 import { text } from "./aktueltText";
 import { fetchData } from "@utils/server/fetch";
-import type { Language } from "../../shared/language/language.ts";
+import type { Language } from "../../utils/server/language.ts";
 import logger from "@utils/server/logger.ts";
 import { Aktuelt } from "./aktueltTypes";
 import { aktueltUrl } from "./aktueltUrls";
