+++ src/features/alert-island/AlertIsland.astro
@@ -1,7 +1,7 @@
 ---
 import Varsler from "./varsler/Varsler.astro";
 import Utkast from "./utkast/Utkast.astro";
-import { getLanguage } from "../../shared/language/language";
+import { getLanguage } from "../../utils/server/language";
 import style from "./AlertIsland.module.css";
 
 const language = getLanguage(Astro.url);
