+++ src/features/alert-island/utkast/Utkast.astro
@@ -1,7 +1,7 @@
 ---
 import UtkastIkon from "./assets/UtkastIkon.tsx";
 import UtkastIkonHover from "./assets/UtkastIkonHover.tsx";
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 import { fetchData } from "@utils/server/fetch";
 import { getAudience, getOboToken } from "@utils/server/token";
 import { utkastApiUrl, utkastUrl } from "./utkastUrls";
