+++ src/features/alert-island/varsler/Varsler.astro
@@ -8,7 +8,7 @@ import { varslerUrl } from "./varslerUrls";
 import IngenVarslerIkon from "./assets/IngenVarslerIkon";
 import VarselBjelleDotMedFyll from "./assets/VarselBjelleDotMedFyll";
 import VarselBjelleDotUtenFyll from "./assets/VarselBjelleDotUtenFyll";
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 import ClientError from "@shared/client-error/ClientError";
 import PersonalContentLogger from "@shared/legacy/PersonalContentLogger";
 import logger from "@utils/server/logger";
