+++ src/features/din-oversikt/DinOversikt.astro
@@ -12,7 +12,7 @@ import { MultiStatusError } from "@utils/server/error";
 import { getAudience, getOboToken } from "@utils/server/token";
 import { PersonalizedContent } from "./DinOversiktTypes";
 import ClientError from "@shared/client-error/ClientError";
-import type { Language } from "../../shared/language/language";
+import type { Language } from "../../utils/server/language";
 import logger from "@utils/server/logger";
 import MeldekortSkeleton from "./meldekort/MeldekortSkeleton.astro";
 import MicrofrontendSSRSkeleton from "../../microfrontends/MicrofrontendSkeleton.astro";
