+++ src/features/dokumenter/Dokumenter.astro
@@ -1,7 +1,7 @@
 ---
 import Dokument from "./dokument/Dokument.astro";
 import ClientError from "@shared/client-error/ClientError";
-import { Language } from "../../shared/language/language";
+import { Language } from "../../utils/server/language";
 import { getAudience, getOboToken } from "../../utils/server/token";
 import { fetchData } from "../../utils/server/fetch";
 import { dokumentarkivUrl, dokumenterUrl } from "./dokumenterUrls";
