+++ src/features/dokumenter/fallback/DokumenterFallback.astro
@@ -1,7 +1,7 @@
 ---
 import { Heading, Skeleton } from "@navikt/ds-react";
 import { text } from "../dokumenterText";
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 
 const language = getLanguage(Astro.url);
 ---
