+++ src/features/dokumenter/ingen/MinIDDokumenter.astro
@@ -1,7 +1,7 @@
 ---
 import { text } from "../dokumenterText";
 import { BodyLong, Heading } from "@navikt/ds-react";
-import { Language } from "../../../shared/language/language";
+import { Language } from "../../../utils/server/language";
 import styles from "../Dokumenter.module.css";
 
 interface Props {
