+++ src/features/innboks/Innboks.astro
@@ -1,6 +1,6 @@
 ---
 import InnboksTag from "./InnboksTag";
-import { Language } from "../../shared/language/language";
+import { Language } from "../../utils/server/language";
 import { innboksUrl } from "./innboksUrls";
 import { antallVarslerUrl } from "./innboksUrls";
 import { text } from "./innboksText";
