+++ src/features/innboks/fallback/InnboksFallback.astro
@@ -1,5 +1,5 @@
 ---
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 import { innboksUrl } from "../innboksUrls";
 import { text } from "../innboksText";
 import { BodyLong, BodyShort, Skeleton } from "@navikt/ds-react";
