+++ src/features/innloggede-tjenester/section/InnloggedeTjenesterSection.astro
@@ -1,7 +1,7 @@
 ---
 import { Detail, BodyShort } from "@navikt/ds-react";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 import LinkWithAmplitude from "../link/LinkWithAmplitude";
 import style from "./InnloggedeTjensterSection.module.css";
