+++ src/features/personalia/Personalia.astro
@@ -1,6 +1,6 @@
 ---
 import { BodyShort } from "@navikt/ds-react";
-import { getLanguage } from "../../shared/language/language";
+import { getLanguage } from "../../utils/server/language";
 import { text } from "./personaliaText";
 import style from "./Personalia.module.css";
 import Navn from "./Navn.astro";
