+++ src/features/substantial-info/SubstantialInfo.astro
@@ -1,7 +1,7 @@
 ---
 import { Alert, Heading, Link, BodyShort } from "@navikt/ds-react";
 import { text } from "./substantialInfoText";
-import type { Language } from "../../shared/language/language";
+import type { Language } from "../../utils/server/language";
 
 interface Props {
   language: Language;
