+++ src/features/utbetaling/Utbetaling.astro
@@ -1,6 +1,6 @@
 ---
 import { getAudience, getOboToken } from "../../utils/server/token";
-import type { Language } from "../../shared/language/language";
+import type { Language } from "../../utils/server/language";
 import IngenUtbetaling from "./ingen/IngenUtbetaling.astro";
 import ClientError from "@shared/client-error/ClientError";
 import { utbetalingsoversiktApiUrl } from "./utbetalingUrls";
