+++ src/features/utbetaling/enkel/EnkelUtbetaling.astro
@@ -3,7 +3,7 @@ import { Heading, BodyShort } from "@navikt/ds-react";
 import { text } from "../utbetalingText";
 import { formatToReadableDate } from "../utbetalingUtils";
 import type { UtbetalingData } from "../utbetalingTypes";
-import type { Language } from "../../../shared/language/language";
+import type { Language } from "../../../utils/server/language";
 import styles from "./EnkelUtbetaling.module.css";
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
