+++ src/features/utbetaling/fallback/UtbetalingFallback.astro
@@ -1,6 +1,6 @@
 ---
 import { Skeleton, Heading } from "@navikt/ds-react";
-import { getLanguage } from "../../../shared/language/language";
+import { getLanguage } from "../../../utils/server/language";
 import { text } from "../utbetalingText";
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import style from "./UtbetalingFallback.module.css";
