+++ src/features/utbetaling/ingen/IngenUtbetaling.astro
@@ -1,6 +1,6 @@
 ---
 import { Heading, BodyShort } from "@navikt/ds-react";
-import { Language } from "../../../shared/language/language";
+import { Language } from "../../../utils/server/language";
 import { text } from "../utbetalingText";
 import styles from "./IngenUtbetaling.module.css";
