+++ src/features/utbetaling/list/UtbetalingList.astro
@@ -1,5 +1,5 @@
 ---
-import { Language } from "../../../shared/language/language";
+import { Language } from "../../../utils/server/language";
 import { UtbetalingData } from "../utbetalingTypes";
 import { BodyShort, Heading, Tag } from "@navikt/ds-react";
 import { formatToReadableDate } from "../utbetalingUtils";
