+++ src/features/utbetaling/se-alle/SeAlle.astro
@@ -1,7 +1,7 @@
 ---
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import { text } from "../utbetalingText";
-import { Language } from "../../../shared/language/language";
+import { Language } from "../../../utils/server/language";
 import styles from "./SeAlle.module.css";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
