+++ src/layouts/base/Layout.astro
@@ -2,7 +2,7 @@
 import Authentication from "@shared/authentication/Authentication";
 import { fetchDecoratorHtml } from "@navikt/nav-dekoratoren-moduler/ssr";
 import { getDecoratorEnvironment } from "../../utils/server/environment";
-import { getLanguage } from "../../shared/language/language";
+import { getLanguage } from "../../utils/server/language";
 import "./Layout.css";
 
 export interface Props {
