+++ src/pages/[lang]/index.astro
@@ -9,7 +9,7 @@ import Container from "@shared/container/Container.astro";
 import Innboks from "@features/innboks/Innboks.astro";
 import Utbetaling from "@features/utbetaling/Utbetaling.astro";
 import Feilmelding from "@shared/feilmelding/Feilmelding";
-import { getLanguage } from "../../shared/language/language";
+import { getLanguage } from "../../utils/server/language";
 import Observability from "@shared/obersvability/Observability";
 import Aktuelt from "../../features/aktuelt/Aktuelt.astro";
 import Dokumenter from "../../features/dokumenter/Dokumenter.astro";
