+++ src/pages/index.astro
@@ -11,7 +11,7 @@ import Utbetaling from "@features/utbetaling/Utbetaling.astro";
 import UtbetalingFallback from "../features/utbetaling/fallback/UtbetalingFallback.astro";
 import Innboks from "@features/innboks/Innboks.astro";
 import Feilmelding from "@shared/feilmelding/Feilmelding";
-import { getLanguage } from "../shared/language/language";
+import { getLanguage } from "../utils/server/language";
 import Observability from "@shared/obersvability/Observability";
 import Aktuelt from "@features/aktuelt/Aktuelt.astro";
 import InnboksFallback from "../features/innboks/fallback/InnboksFallback.astro";
